Move column components to features
