+++ src/pages/boards/[id].tsx
@@ -4,8 +4,8 @@ import { useRouter } from "next/router";
 import styled from "styled-components";
 import Image from "next/image";
 import { Navbar } from "src/features/board/navbar";
-import { NewColumn } from "src/components/columns/new-column";
-import { ColumnList } from "src/components/columns/column-list";
+import { NewColumn } from "src/features/column/new-column";
+import { ColumnList } from "src/features/column/column-list";
 import { Subnav } from "src/features/board/subnav";
 import { imagePath } from "src/utils/image-paths";
 import { getBoardThunk, selectBoardById } from "src/redux/board.slice";
